fix: Move tests to pre-existing folder, following existing structure
- Removed integration tests since they were completely unused
- Pre-existing governance tests have been moved to
  /tests/src/governance/proposals
- Payment-related tests have been merged and moved to
  /tests/src/governance/payments
- Some payment-related tests were either unused or unfinished, so I
  completed them. All tests should be accurate now.
